refactor: date 함수 공용 유틸로 사용
