+++ src/page-layer/attendance/utils/formatter.ts
@@ -1,16 +1,5 @@
 import { parseDateString as sharedParseDateString } from "@/shared/utils";
 
-export const formatDateToString = (date: Date): string => {
-  const year = date.getFullYear();
-  const month = String(date.getMonth() + 1).padStart(2, "0");
-  const day = String(date.getDate()).padStart(2, "0");
-  return `${year}-${month}-${day}`;
-};
-
-export const formatDatesToStrings = (dates: Date[]): string[] => {
-  return dates.map(formatDateToString);
-};
-
 /**
  * @deprecated Use parseDateString from @/shared/utils instead
  */
