+++ src/shared/utils/formatter.ts
@@ -21,3 +21,14 @@ export const parseDateString = (dateString: string): Date => {
   const [year, month, day] = dateString.split("-").map(Number);
   return new Date(year, month - 1, day);
 };
+
+export const formatDateToString = (date: Date): string => {
+  const year = date.getFullYear();
+  const month = String(date.getMonth() + 1).padStart(2, "0");
+  const day = String(date.getDate()).padStart(2, "0");
+  return `${year}-${month}-${day}`;
+};
+
+export const formatDatesToStrings = (dates: Date[]): string[] => {
+  return dates.map(formatDateToString);
+};
